sprint 12/1 p2p review

+++ src/main/java/ru/yandex/practicum/filmorate/exception/UserNotFoundException.java
@@ -2,7 +2,7 @@ package ru.yandex.practicum.filmorate.exception;
 
 public class UserNotFoundException extends NotFoundException {
 
-    public UserNotFoundException(Integer id) {
+    public UserNotFoundException(long id) {
         super(String.format("Пользователь с id=%s не найден", id));
     }
 }
